% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prevalence.R
\name{desc_prevalence}
\alias{desc_prevalence}
\title{Describe the prevalence of a disease/disorder.}
\usage{
desc_prevalence(.data, .variables)
}
\arguments{
\item{.data}{Dataset with SID, VN, and the variable(s) of interest.}

\item{.variables}{Columns to apply the function to.}
}
\value{
Wide dataset with time across the top and various disease/disorder
statuses below.
}
\description{
Input must be long format with subject id (SID), visit number (VN),
and the column(s) to determine the prevalence.
}
\examples{

example_data \%>\%
desc_prevalence("DM")

}
