% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{desc_table}
\alias{desc_table}
\title{Create a table from the data frame.}
\usage{
desc_table(.data, caption = NULL)
}
\arguments{
\item{.data}{Dataset to create into a table.}

\item{caption}{Table caption.}
}
\value{
Output the results of some descriptions of the dataset.
}
\description{
Simply a thin wrapper around [pander::pander].
}
\examples{

\dontrun{
example_data \%>\%
desc_prevalence("DM") \%>\%
desc_table()
}

}
