% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PROMISE.desc-pkg.R
\docType{data}
\name{example_data}
\alias{example_data}
\title{Example dataset with disease status for subjects over time.}
\format{A data frame with 1131 observations and 5 variables:
\itemize{
\item SID. Subject identification number. Each observation is a specific subject at a specific time.
\item VN. Visit number, or the visit year when the subject came in for data collection.
\item DM. Diabetes mellitus status (0 = does not have, 1 = does have).
\item IFG. Impaired fasting glucose (0 = not have, 1 = has).
\item IGT. Impaired glucose tolerance (0 = not have, 1 = has).
}}
\source{
Subsetted and sampled (with replacement) from the PROMISE dataset.
http://promise-cohort.gitlab.io/PROMISE
}
\usage{
example_data
}
\description{
This dataset is a modified version of the PROMISE dataset to highlight how to
use this package.
}
\keyword{datasets}
