% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PROMISE.desc-pkg.R
\docType{package}
\name{PROMISE.desc-package}
\alias{PROMISE.desc}
\alias{PROMISE.desc-package}
\title{PROMISE.desc: Describe Common Statistical Summaries for the PROMISE cohort}
\description{
Calculate and describe some standard summaries for the PROMISE
cohort, such as incidence of diabetes or dysglycemia or change over time in
variables.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/lwjohnst86/PROMISE.desc}
  \item Report bugs at \url{https://github.com/lwjohnst86/PROMISE.desc/issues}
}

}
\author{
\strong{Maintainer}: Luke Johnston \email{lwjohnst@gmail.com}

}
\keyword{internal}
